Roles and authorization

+++ src/main/java/com/allianceever/projectERP/controller/EmployeeController.java
@@ -8,6 +8,8 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.BeanWrapper;
 import org.springframework.beans.BeanWrapperImpl;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -39,10 +41,18 @@ public class EmployeeController {
 
     // Build Get Employee REST API
     @GetMapping("/{id}")
-    public ResponseEntity<EmployeeDto> getEmployeeById(@PathVariable("id") Long employeeID){
+    public ResponseEntity<EmployeeDto> getEmployeeById(@PathVariable("id") Long employeeID, @AuthenticationPrincipal Jwt jwt){
         EmployeeDto employeeDto = employeeService.getById(employeeID);
         if (employeeDto != null) {
-            return ResponseEntity.ok(employeeDto);
+            // Retrieve username and role from the jwt
+            String username = jwt.getClaimAsString("sub");
+            String role = jwt.getClaimAsString("roles");
+
+            if(employeeDto.getUserName().equals(username) || role.equals("ADMIN") || role.equals("Human_Capital")){
+                return ResponseEntity.ok(employeeDto);
+            }else{
+                return ResponseEntity.notFound().build();
+            }
         } else {
             return ResponseEntity.notFound().build();
         }
@@ -58,7 +68,11 @@ public class EmployeeController {
     // Build Multipart Update Employee REST API
     @SneakyThrows
     @PostMapping("/updateEmployeeMultipart")
-    public ResponseEntity<EmployeeDto> updateEmployeeMultipart(@ModelAttribute EmployeeDto employeeDto, @RequestParam("imageFile") MultipartFile imageFile){
+    public ResponseEntity<EmployeeDto> updateEmployeeMultipart(@ModelAttribute EmployeeDto employeeDto, @RequestParam("imageFile") MultipartFile imageFile, @AuthenticationPrincipal Jwt jwt){
+        // Retrieve username and role from the jwt Token
+        String username = jwt.getClaimAsString("sub");
+        String role = jwt.getClaimAsString("roles");
+
         String uploadDir = "./src/main/resources/static/assets/img/profiles/";
         Long EmployeeID = employeeDto.getEmployeeID();
         EmployeeDto existingEmployee = employeeService.getById(EmployeeID);
@@ -66,27 +80,30 @@ public class EmployeeController {
         if (existingEmployee == null) {
             return ResponseEntity.notFound().build();
         }
+        if(existingEmployee.getUserName().equals(username) || role.equals("ADMIN") || role.equals("Human_Capital")){
+            if (imageFile != null && !imageFile.isEmpty()) {
+                String fileName = "imgP" + EmployeeID + "." + StringUtils.getFilenameExtension(imageFile.getOriginalFilename());
+                try {
+                    // Save the file in the specified upload directory
+                    Path filePath = Paths.get(uploadDir, fileName);
+                    Files.copy(imageFile.getInputStream(), filePath, StandardCopyOption.REPLACE_EXISTING);
 
-        if (imageFile != null && !imageFile.isEmpty()) {
-            String fileName = "imgP" + EmployeeID + "." + StringUtils.getFilenameExtension(imageFile.getOriginalFilename());
-            try {
-                // Save the file in the specified upload directory
-                Path filePath = Paths.get(uploadDir, fileName);
-                Files.copy(imageFile.getInputStream(), filePath, StandardCopyOption.REPLACE_EXISTING);
-
-                // Save the fileName in the database (you need to add a field for image name in your database)
-                existingEmployee.setImageName(fileName);
-            } catch (IOException e) {
-                e.printStackTrace();
+                    // Save the fileName in the database (you need to add a field for image name in your database)
+                    existingEmployee.setImageName(fileName);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
+
+            // Perform a partial update of the existingEmployee using the employeeDto data
+            BeanUtils.copyProperties(employeeDto, existingEmployee, getNullPropertyNames(employeeDto));
+
+            // Save the updated employee data back to the database
+            EmployeeDto updatedEmployee = employeeService.update(EmployeeID,existingEmployee);
+            return ResponseEntity.ok(updatedEmployee);
+        }else{
+            return ResponseEntity.notFound().build();
         }
-
-        // Perform a partial update of the existingEmployee using the employeeDto data
-        BeanUtils.copyProperties(employeeDto, existingEmployee, getNullPropertyNames(employeeDto));
-
-        // Save the updated employee data back to the database
-        EmployeeDto updatedEmployee = employeeService.update(EmployeeID,existingEmployee);
-        return ResponseEntity.ok(updatedEmployee);
     }
 
     // Build Update Employee REST API
