+++ src/main/java/com/allianceever/projectERP/controller/LeaveTypeController.java
@@ -6,6 +6,8 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
@@ -25,19 +27,31 @@ public class LeaveTypeController {
     }
 
     @PostMapping("/create")
-    public ResponseEntity<LeaveTypeDto> createLeaveType(@ModelAttribute() LeaveTypeDto leaveTypeDto){
+    public ResponseEntity<LeaveTypeDto> createLeaveType(@ModelAttribute() LeaveTypeDto leaveTypeDto, @AuthenticationPrincipal Jwt jwt){
+        // Retrieve username from the jwt
+        String username = jwt.getClaimAsString("sub");
+
+        String leaveName = leaveTypeDto.getLeaveName()+"("+username+")";
+
+        leaveTypeDto.setUsername(username);
+        leaveTypeDto.setLeaveName(leaveName);
         LeaveTypeDto createdLeaveType = leaveTypeService.create(leaveTypeDto);
         return new ResponseEntity<>(createdLeaveType, CREATED);
     }
 
     @GetMapping("/{leaveTypeName}")
-    public ResponseEntity<LeaveTypeDto> getHolidayByHolidayName(@PathVariable("leaveTypeName") String leaveTypeName){
+    public ResponseEntity<LeaveTypeDto> getHolidayByHolidayName(@PathVariable("leaveTypeName") String leaveTypeName, @AuthenticationPrincipal Jwt jwt){
+        // Retrieve username and role from the jwt
+        String username = jwt.getClaimAsString("sub");
+        String role = jwt.getClaimAsString("roles");
+
         LeaveTypeDto leaveTypeDto = leaveTypeService.getByLeaveName(leaveTypeName);
         if (leaveTypeDto != null) {
-            return ResponseEntity.ok(leaveTypeDto);
-        } else {
-            return ResponseEntity.notFound().build();
+            if(leaveTypeDto.getUsername().equals(username) || role.equals("ADMIN") || role.equals("Human_Capital")){
+                return ResponseEntity.ok(leaveTypeDto);
+            }
         }
+        return ResponseEntity.notFound().build();
     }
 
 
@@ -56,20 +70,24 @@ public class LeaveTypeController {
         return ResponseEntity.ok(updatedLeaveType);
     }
 
-    @Transactional
-    // Build Delete Employee REST API
-    @PostMapping("/delete/{LeaveTypeName}")
-    public ResponseEntity<String> deleteLeaveType(@PathVariable("LeaveTypeName")  String leaveName){
-        LeaveTypeDto leaveTypeDto = leaveTypeService.getByLeaveName(leaveName);
-        // if (employeeDto != null) {
-        //   employeeService.delete(employeeID);
-        //   return ResponseEntity.ok("Employee deleted successfully!");
-        //  } else {
-        //    return ResponseEntity.notFound().build();
-        //  }
 
-        leaveTypeService.delete(leaveName);
-        return ResponseEntity.ok("LeaveType deleted successfully!");
+    // Build Delete Employee REST API
+    @Transactional
+    @PostMapping("/delete/{LeaveTypeName}")
+    public ResponseEntity<String> deleteLeaveType(@PathVariable("LeaveTypeName")  String leaveName, @AuthenticationPrincipal Jwt jwt){
+        // Retrieve username and role from the jwt
+        String username = jwt.getClaimAsString("sub");
+        String role = jwt.getClaimAsString("roles");
+
+        LeaveTypeDto leaveTypeDto = leaveTypeService.getByLeaveName(leaveName);
+
+        if (leaveTypeDto != null) {
+            if(leaveTypeDto.getUsername().equals(username) || role.equals("ADMIN") || role.equals("Human_Capital")){
+                leaveTypeService.delete(leaveName);
+                return ResponseEntity.ok("LeaveType deleted successfully!");
+            }
+        }
+        return ResponseEntity.notFound().build();
     }
